generate bar chart for daily activity, rename group to myinfo
