+++ main.py
@@ -27,7 +27,7 @@ from commands.speak import speak, speak_embed
 from commands.trekduel import trekduel
 from commands.trektalk import trektalk
 from commands.tuvix import tuvix
-from commands.userinfo import top_channels
+from commands.myinfo import top_channels
 
 # Slash Command Groups
 import commands.birthday
